VNPay Schedule fixed Feedback fixed Remake some DTO Remake some Services Add additional function to repository

+++ src/main/java/com/swp391/webapp/Repository/PackageRepository.java
@@ -2,7 +2,6 @@ package com.swp391.webapp.Repository;
 
 import com.swp391.webapp.Entity.AccountEntity;
 import com.swp391.webapp.Entity.PackageEntity;
-import com.swp391.webapp.dto.Package;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -14,8 +13,9 @@ public interface PackageRepository extends JpaRepository<PackageEntity, Integer>
 
     List<PackageEntity> findPackagesByAccount(AccountEntity account);
 
-    PackageEntity findPackageByPackageID(long id);
+    PackageEntity findPackageByPackageID(int id);
 
     List<PackageEntity> findPackagesByAccountAndIsDeletedFalse(AccountEntity account);
+
 }
